feat: update method to charge and discharge and minor changes

+++ api.py
@@ -1,37 +1,35 @@
 import logging
 
 from flask import Flask, request, jsonify
-from threading import Lock
 from uuid import uuid4
+from datetime import datetime
 
 from database.db import Session, configure_database
 from database.models import Battery
-from sqlalchemy.orm.exc import NoResultFound
+
+from src.octave_batteries import OctaveBattery
 
 app = Flask(__name__)
 
 logger = logging.getLogger(__name__)
 
-# Thread-safety transaction
-batteries_lock = Lock()
-
 @app.route("/<battery_id>", methods=["GET"])
 def get_battery(battery_id):
     try:
+        # TODO: can create decorators to start and terminate session
         session = Session()
 
         query = session.query(Battery)
-        battery = query.filter_by(battery_id=battery_id).one()
+        battery = query.filter_by(battery_id=battery_id).one_or_none()
+        if battery is None:
+            logger.error(f"error: {e}, status code: 404")
+            return jsonify({"error": "Battery instance not found"}), 404
 
         return battery.to_dict(), 200
 
-    except NoResultFound as e:
-        logger.error(f"error: {e}, status code: 404")
-        return jsonify({"error": "Battery instance not found"}), 404
-
     except Exception as e:
-        logger.error(f"error: {e}, status code: 500")
-        return jsonify({"error": str(e)}), 500
+        logger.error(f"Internal Server Error: {e}, status code: 500")
+        return jsonify({"Internal Server Error": str(e)}), 500
 
     finally:
         session.close()
@@ -55,9 +53,8 @@ def create_battery():
             maximum_power_kw= max_power
         )
 
-        with batteries_lock:
-            session.add(battery)
-            session.commit()
+        session.add(battery)
+        session.commit()
 
         return jsonify({
             "capacity_kwh": battery.capacity_kwh,
@@ -67,8 +64,8 @@ def create_battery():
 
     except Exception as e:
         session.rollback()
-        logger.error(f"error: {e}, status code: 500")
-        return jsonify({"error": str(e)}), 500
+        logger.error(f"Internal Server Error: {e}, status code: 500")
+        return jsonify({"Internal Server Error": str(e)}), 500
 
     finally:
         session.close()
@@ -79,26 +76,71 @@ def delete_battery(battery_id):
         session = Session()
 
         query = session.query(Battery)
-        battery = query.filter_by(battery_id=battery_id).one()
+        battery = query.filter_by(battery_id=battery_id).one_or_none()
 
-        with batteries_lock:
-            session.delete(battery)
-            session.commit()
+        if battery is None:
+            logger.error(f"error: Could not find the battery with ID: {battery_id}, status code: 404")
+            return jsonify({"error": f"Could not find the battery with ID: {battery_id}"}), 404
+
+        session.delete(battery)
+        session.commit()
 
         # Could have used 204 (No Content) but want to return additional response message.
         return jsonify({"message": f"Battery instance with ID: {battery_id} deleted successfully"}), 200
 
-    except NoResultFound as e:
-        logger.error(f"error: Could not find the battery with ID: {battery_id}, status code: 404")
-        return jsonify({"error": f"Could not find the battery with ID: {battery_id}"}), 404
-
     except Exception as e:
         session.rollback()
-        logger.error(f"error: {e}, status code: 500")
-        return jsonify({"error": str(e)}), 500
+        logger.error(f"Internal Server Error: {e}, status code: 500")
+        return jsonify({"Internal Server Error": str(e)}), 500
 
     finally:
         session.close()
+
+# /update?battery_id=<id>&power=<int>&duration=<float>
+@app.route("/update", methods=["PATCH"])
+def update_battery():
+    battery_id = request.args.get('battery_id', type=str)
+    power = request.args.get('power', type=int)
+    duration = request.args.get('duration', type=float)
+
+    if battery_id is None or power is None or duration is None:
+        logger.error("Missing battery_id, power or duration, status code: 400")
+        return jsonify({"error": "Missing battery_id, power or duration"}), 400
+
+    try:
+        session = Session()
+
+        query = session.query(Battery)
+        battery_details = query.filter_by(battery_id=battery_id).one_or_none()
+
+        print("battery_details", battery_details)
+
+        if battery_details:
+            if power > 0:
+                battery_details = OctaveBattery.charge(battery_details, power, duration)
+            elif power < 0:
+                battery_details = OctaveBattery.discharge(battery_details, power, duration)
+
+
+            warning = OctaveBattery.get_warning(battery_details)
+            if warning:
+                print("Warning: ", warning)
+                print("timestamp: ", datetime.now().isoformat())
+                # OctaveBattery.publish_warning(battery_details, warning)
+
+            session.merge(battery_details)
+            session.commit()
+
+            return battery_details.to_dict(), 200
+        else:
+            logger.error(f"error: Could not find the battery with ID: {battery_id}, status code: 404")
+            return jsonify({"error": f"Could not find the battery with ID: {battery_id}"}), 404
+
+
+    except Exception as e:
+        session.rollback()
+        logger.error(f"Internal Server Error: {e}, status code: 500")
+        return jsonify({"Internal Server Error": str(e)}), 500
 
 if __name__ == "__main__":
     configure_database()
